+++ pkg/ui/react-app/src/pages/graph/Panel.test.tsx
@@ -22,6 +22,8 @@ const defaultProps: PanelProps = {
     usePartialResponse: false,
     storeMatches: [],
     engine: 'prometheus',
+    explain: false,
+    disableExplainCheckbox: false,
   },
   onOptionsChanged: (): void => {
     // Do nothing.
@@ -55,6 +57,9 @@ describe('Panel', () => {
       results.push(opts);
     };
     const panel = shallow(<Panel {...defaultProps} onOptionsChanged={onOptionsChanged} />);
+    // Panel construction updates Explain checkbox prop to disbale.
+    // Hence, a result is added and dropping it.
+    results.length = 0;
     const links = panel.find(NavLink);
     [
       { panelType: 'Table', active: true },
@@ -92,6 +97,8 @@ describe('Panel', () => {
       usePartialResponse: false,
       storeMatches: [],
       engine: 'prometheus',
+      explain: false,
+      disableExplainCheckbox: false,
     };
     const graphPanel = mount(<Panel {...defaultProps} options={options} />);
     const controls = graphPanel.find(GraphControls);